test: migrate recommended_app_service tests to testcontainers
